Add form partial for create and edit template

+++ src/Controller/PinsController.php
@@ -9,9 +9,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Entity\Pin;
 use App\Form\PinType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+
 
 class PinsController extends AbstractController
 {
@@ -26,7 +25,7 @@ class PinsController extends AbstractController
     public function create(Request $request, EntityManagerInterface $em): Response
     {
         $pin = new Pin;
-        $form = $this->createForm(PinType::class,$pin);
+        $form = $this->createForm(PinType::class, $pin);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $em->persist($pin);
@@ -47,7 +46,7 @@ class PinsController extends AbstractController
     #[Route("/pins/{id<\d+>}/edit", name: 'app_pin_edit')]
     public function edit(Pin $pin, Request $request, EntityManagerInterface $em): Response
     {
-        $form = $this->createForm(PinType::class,$pin);
+        $form = $this->createForm(PinType::class, $pin);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
@@ -59,10 +58,14 @@ class PinsController extends AbstractController
         return $this->render('pins/edit.html.twig', ['pin' => $pin, 'form' => $form->createView()]);
     }
     #[Route("/pins/{id<\d+>}/delete", name: 'app_pin_delete')]
-    public function delete(Pin $pin,EntityManagerInterface $em): Response
+    public function delete(Pin $pin, EntityManagerInterface $em, Request $request): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        $csrf = $request->request->get('csrf_token');
+        if ($this->isCsrfTokenValid('pin_deletion' . $pin->getId(), $csrf)) {
+            $em->remove($pin);
+            $em->flush();
+        }
+
         return $this->redirectToRoute('app_home');
     }
 }
